requirements.txt
pandas==2.2.2
openai==1.52.0
python-dotenv==1.0.1
httpx<0.28.0
numpy>=1.26.0
xai-sdk>=0.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
google-cloud-bigquery
db-dtypes
streamlit
google-cloud-bigquery-storage
anthropic
pandas_gbq
